feat: Refactor LandingPage and establish feature structure - Decomposed the LandingPage into smaller, reusable components. - Organized components into a feature-sliced directory structure (employer, task, layout).
