+++ frontend/workerfrontend/src/App.tsx
@@ -1,5 +1,4 @@
-import React from "react";
-import LandingPage from "./components/LandingPage";
+import LandingPage from "./pages/LandingPage";
 
 export default function App() {
   return <LandingPage />;
