+++ frontend/workerfrontend/src/features/employer/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export * from './components/EmployerPromoCard';
+
