+++ frontend/workerfrontend/src/features/task/index.ts
@@ -1,0 +1,5 @@
+export * from './types';
+export * from './components/TaskCard';
+export * from './components/TaskerPromoCard';
+export * from './components/TaskFilter';
+export * from './components/TaskList';
