+++ frontend/workerfrontend/src/features/task/types.ts
@@ -1,0 +1,10 @@
+export type Kategoria = "cleaning" | "garden" | "tech" | "pets" | "vehicles" | "all";
+
+export interface Task {
+  id: string;
+  otsikko: string;
+  kategoria: Kategoria;
+  hinta: string;
+  sijainti: string;
+  pvm: string;
+}
